+++ coop_account/__manifest__.py
@@ -12,10 +12,13 @@
     'license': 'AGPL-3',
     'depends': [
         'account',
-            'queue_job'
+        'queue_job',
+        'barcodes_generator_partner',
+        'account_tax_balance',
     ],
     'data': [
         "security/res_group.xml",
+        "security/ir.model.access.csv",
         "view/assets.xml",
         "view/view_account_bank_statement.xml",
         "view/view_account_move.xml",
@@ -25,7 +28,6 @@
         "view/view_account_invoice.xml",
         "wizard/view_bank_statement_line_reconcile_wizard.xml",
         "wizard/view_unmatch_bank_statement_wizard.xml",
-
         "view/menu.xml",
     ],
     'installable': True,
