+++ coop_account/security/res_group.xml
@@ -7,4 +7,8 @@
         <field name="implied_ids" eval="[(4, ref('account.group_account_manager'))]"/>
     </record>
 
+    <record id="group_validate_invoice" model="res.groups">
+        <field name="name">Validate Invoices</field>
+    </record>
+
 </odoo>
