+++ coop_account/tests/__init__.py
@@ -1,0 +1,2 @@
+from . import common
+from . import test_coop_custom_invoice
