+++ coop_account/view/view_account_payment.xml
@@ -78,6 +78,12 @@
             <field name="company_id" position="before">
                 <field name="invoice_ids" />
             </field>
+            <field name="partner_id" position="after">
+                <field name="communication" />
+            </field>
+            <field name="payment_date" position="before">
+                <field name="partner_code" />
+            </field>
         </field>
     </record>
 
@@ -108,4 +114,36 @@
               parent="account.menu_finance_entries"
               sequence="25"/>
 
+    <record id="view_account_payment_form" model="ir.ui.view">
+        <field name="model">account.payment</field>
+        <field name="inherit_id" ref="account.view_account_payment_form"/>
+        <field name="arch" type="xml">
+            <field name="communication" position="attributes">
+                <attribute name="attrs">{
+                    'invisible': [('state', '!=', 'draft'), ('communication', '=', False)],
+                    'readonly': [('state', '!=', 'draft')],
+                    'required': [('payment_type', '=', 'transfer')]}
+                </attribute>
+            </field>
+            <field name="communication" position="before">
+                <field name="operation_type" />
+                <field name="text_check_code"
+                    attrs="{'invisible': [('operation_type', '!=', 'check')],
+                    'required': [('operation_type', '=', 'check')]}"/>
+                <field name="text_lcr_code"
+                    attrs="{'invisible': [('operation_type', '!=', 'lcr')],
+                    'required': [('operation_type', '=', 'lcr')]}"/>
+            </field>
+        </field>
+    </record>
+
+    <record id="view_account_payment_search" model="ir.ui.view">
+        <field name="model">account.payment</field>
+        <field name="inherit_id" ref="account.view_account_payment_search"/>
+        <field name="arch" type="xml">
+            <field name="partner_id" position="after">
+                <field name="partner_code" />
+            </field>
+        </field>
+    </record>
 </odoo>
