--- louve_custom_account/__init__.py
@@ -1,7 +1,0 @@
-# Copyright (C) 2017-Today: La Louve (<http://www.lalouve.net/>)
-# Copyright (C) 2019-Today: Druidoo (<https://www.druidoo.io>)
-# @author: Sylvain LE GAL (https://twitter.com/legalsylvain)
-#          Julien Weste
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
-from . import models
