--- louve_custom_account/i18n/fr.po
@@ -1,115 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-#	* louve_custom_account
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 12.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2020-02-20 06:44+0000\n"
-"PO-Revision-Date: 2020-02-20 06:44+0000\n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: louve_custom_account
-#: model:ir.model.fields,field_description:louve_custom_account.field_account_payment__partner_code
-msgid "Barcode Base"
-msgstr "Base du code Barre"
-
-#. module: louve_custom_account
-#: selection:account.payment,operation_type:0
-msgid "Check"
-msgstr "Chèque"
-
-#. module: louve_custom_account
-#: model:ir.model.fields,field_description:louve_custom_account.field_account_payment__text_check_code
-msgid "Check Code"
-msgstr "Vérifier le code"
-
-#. module: louve_custom_account
-#: code:addons/louve_custom_account/models/account_payment.py:54
-#, python-format
-msgid "Check nb "
-msgstr "CHQ"
-
-#. module: louve_custom_account
-#: code:addons/louve_custom_account/models/account_payment.py:43
-#, python-format
-msgid "Credit Card "
-msgstr "CB"
-
-#. module: louve_custom_account
-#: selection:account.payment,operation_type:0
-msgid "Credit card"
-msgstr "Carte bancaire"
-
-#. module: louve_custom_account
-#: model:ir.model,name:louve_custom_account.model_account_invoice
-msgid "Invoice"
-msgstr "Facture"
-
-#. module: louve_custom_account
-#: selection:account.payment,operation_type:0
-msgid "LCR"
-msgstr ""
-
-#. module: louve_custom_account
-#: model:ir.model.fields,field_description:louve_custom_account.field_account_payment__text_lcr_code
-msgid "LCR Code"
-msgstr ""
-
-#. module: louve_custom_account
-#: code:addons/louve_custom_account/models/account_payment.py:58
-#, python-format
-msgid "LCR nb "
-msgstr "LCR "
-
-#. module: louve_custom_account
-#: model:ir.actions.server,name:louve_custom_account.action_server_merge_lines
-msgid "Merge Invoice Lines"
-msgstr "Fusionner les lignes de facture"
-
-#. module: louve_custom_account
-#: model:ir.model.fields,field_description:louve_custom_account.field_account_payment__operation_type
-msgid "Operation Type"
-msgstr "Type d'opération"
-
-#. module: louve_custom_account
-#: selection:account.payment,operation_type:0
-msgid "Other"
-msgstr "Autre"
-
-#. module: louve_custom_account
-#: model:ir.model,name:louve_custom_account.model_account_payment
-msgid "Payments"
-msgstr "Paiements"
-
-#. module: louve_custom_account
-#: selection:account.payment,operation_type:0
-#: code:addons/louve_custom_account/models/account_payment.py:40
-#, python-format
-msgid "SEPA Direct Credit"
-msgstr "Virement SEPA"
-
-#. module: louve_custom_account
-#: selection:account.payment,operation_type:0
-#: code:addons/louve_custom_account/models/account_payment.py:37
-#, python-format
-msgid "SEPA Direct Debit"
-msgstr "Prélèvement SEPA"
-
-#. module: louve_custom_account
-#: model:res.groups,name:louve_custom_account.group_validate_invoice
-msgid "Validate Invoices"
-msgstr "Valider les factures"
-
-#. module: louve_custom_account
-#: code:addons/louve_custom_account/models/account_invoice.py:14
-#, python-format
-msgid "You can only merge draft invoice!"
-msgstr "Vous ne pouvez fusionner que le projet de facture!"
-
