--- louve_custom_account/models/__init__.py
@@ -1,3 +1,0 @@
-
-from . import account_payment
-from . import account_invoice
