--- louve_custom_account/security/res_group.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <record id="group_validate_invoice" model="res.groups">
-        <field name="name">Validate Invoices</field>
-    </record>
-
-</odoo>
