--- louve_custom_account/tests/__init__.py
@@ -1,2 +1,0 @@
-from . import common
-from . import test_louve_custom_invoice
