--- louve_custom_account/views/view_account_move_line.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0"?>
-<odoo>
-
-    <record id="view_account_move_line_inherit" model="ir.ui.view">
-        <field name="model">account.move.line</field>
-        <field name="inherit_id" ref="account.view_move_line_tree"/>
-        <field name="arch" type="xml">
-            <field name="analytic_account_id" position="after">
-                <field name="tax_ids" widget="many2many_tags"/>
-            </field>
-        </field>
-    </record>
-
-</odoo>
